refactor navbar code remove repetitive code
